requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==23.0.0
packaging==25.0
psycopg2-binary==2.9.11
sqlparse==0.5.4
typing_extensions==4.15.0